Move Slack command tests into separate files
