+++ api/tests/slack/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Blossom Slack commands."""
